Cambio para realizar la media con decimales

+++ src/main/java/com/example/back_simulador_comunio/entities/Jugador.java
@@ -19,7 +19,7 @@ public class Jugador {
 	private String nombre;
 	
 	@Column(name="puntos_totales")
-	private Integer puntosTotales;
+	private Float puntosTotales;
 	
 	@Column(name="puntos_media")
 	private Float puntosMedia;
@@ -69,11 +69,11 @@ public class Jugador {
 		this.nombre = nombre;
 	}
 
-	public Integer getPuntosTotales() {
+	public Float getPuntosTotales() {
 		return puntosTotales;
 	}
 
-	public void setPuntosTotales(Integer puntosTotales) {
+	public void setPuntosTotales(Float puntosTotales) {
 		this.puntosTotales = puntosTotales;
 	}
